serverless/es-curator/requirements.txt
boto3==1.9.66
botocore==1.12.66
certifi==2018.11.29
chardet==3.0.4
click==6.7
DateTime==4.3
docutils==0.14
elasticsearch==6.3.1
elasticsearch-curator==5.6.0
futures==3.2.0
idna==2.8
jmespath==0.9.3
pybase64==0.4.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.21.0
requests-aws4auth==0.9
s3transfer==0.1.13
six==1.12.0
slackclient==1.3.0
urllib3==1.24.1
voluptuous==0.11.5
websocket-client==0.54.0
zope.interface==4.6.0
